refactor: restructure nvim config into modular plugin files
- Move all plugin specs from monolithic lazy.lua into lua/klepp0/plugins/
- Each plugin file is self-contained: declaration, config, and keymaps
- Remove after/plugin/ directory entirely (13 files)
- Remove duplicate plugin declarations (nvim-dap, harpoon)
- Remove dead which-key config
- Remove commented-out kickstart.nvim boilerplate from init.lua
- Remove conflicting <leader>f mapping from remap.lua (conform owns it)
- Slim lazy.lua from 635 lines to ~25 (bootstrap only)
