--- after/plugin/harpoon.lua
@@ -1,37 +1,0 @@
-local mark = require("harpoon.mark")
-local ui = require("harpoon.ui")
-
-vim.keymap.set("n", "<C-a>", mark.add_file, { desc = "[A]ppend File to Harpoon" })
-vim.keymap.set("n", "<leader>c", mark.clear_all, { desc = "[C]lear Harpoon" })
-vim.keymap.set("n", "<C-e>", ui.toggle_quick_menu, { desc = "Open Harpoon [E]xplorer" })
-
-vim.keymap.set("n", "<leader>1", function()
-	ui.nav_file(1)
-end, { desc = "Jump to Mark [1]" })
-vim.keymap.set("n", "<leader>2", function()
-	ui.nav_file(2)
-end, { desc = "Jump to Mark [2]" })
-vim.keymap.set("n", "<leader>3", function()
-	ui.nav_file(3)
-end, { desc = "Jump to Mark [3]" })
-vim.keymap.set("n", "<leader>4", function()
-	ui.nav_file(4)
-end, { desc = "Jump to Mark [4]" })
-vim.keymap.set("n", "<leader>5", function()
-	ui.nav_file(5)
-end, { desc = "Jump to Mark [5]" })
-vim.keymap.set("n", "<leader>6", function()
-	ui.nav_file(6)
-end, { desc = "Jump to Mark [6]" })
-vim.keymap.set("n", "<leader>7", function()
-	ui.nav_file(7)
-end, { desc = "Jump to Mark [7]" })
-vim.keymap.set("n", "<leader>8", function()
-	ui.nav_file(8)
-end, { desc = "Jump to Mark [8]" })
-vim.keymap.set("n", "<leader>9", function()
-	ui.nav_file(9)
-end, { desc = "Jump to Mark [9]" })
-vim.keymap.set("n", "<leader>0", function()
-	ui.nav_file(0)
-end, { desc = "Jump to Mark [10]" })
