+++ lua/klepp0/plugins/colorscheme.lua
@@ -1,0 +1,8 @@
+return {
+	"rose-pine/neovim",
+	priority = 1000,
+	init = function()
+		vim.cmd.colorscheme("rose-pine-main")
+		vim.cmd.hi("Comment gui=none")
+	end,
+}
